Refactor: dependencyInjector 수정
