+++ server/loaders/dependencyInjector.js
@@ -1,21 +1,24 @@
 const { Container } = require("typedi");
+const dependencyConvert = require("./dependencyConvert");
 
 /** typedi Container에 의존성들을 주입하는 함수
  * @description Container.get으로 불러오기 전에 실행되어야 함
- * @param {Object} container - 컨테이너 객체
- * @param {Array.<Class>} container.models - 사용할 모델 클래스들의 배열.
- * @param {Array.<Class>} container.services - 사용할 서비스 클래스들의 배열.
+ * @param {Object.<Class>} container - 컨테이너 객체
+ * @param {Object.<Class>} container.models - 사용할 모델 클래스들의 객체.
+ * @param {Object} container.services - 사용할 서비스 클래스들 객체.
  * @return {void} 의존성 주입
  */
 module.exports = ({ models, services }) => {
+  const { modelsArr, servicesArr } = dependencyConvert(models, services);
+
   try {
-    models.forEach((m) => {
+    modelsArr.forEach((m) => {
       console.log(m);
       Container.set({ id: m.name, factory: () => m.model });
       console.log(`👌 ${m.name} injected into container`);
     });
 
-    services.forEach((s) => {
+    servicesArr.forEach((s) => {
       console.log(s);
       Container.set({ id: s.name, factory: () => new s.service() });
       console.log(`👌 ${s.name} injected into container`);
